<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation


main:

```shell
Skip N=576, K=7168 now
deepseek-ai/DeepSeek-V3 N=24576 K=7168: 
fp8 blockwise scaled matmul:
    batch_size         vllm   sgl-kernel  sglang triton     deepgemm
0          1.0    89.951999    93.024001      84.895998    76.544002
1          8.0    81.216000    84.063999      70.303999    70.175998
2         16.0    82.336001    85.023999      76.223999    64.640000
3         32.0    80.991998    83.552003      64.000003    59.136000
4         64.0    82.064003    84.991999      62.944002    57.760000
5        128.0    77.087998    80.031998      97.952001    61.152000
6        256.0   105.343997   107.391998     143.040001    87.296002
7        512.0   199.167997   197.104007     271.295995   138.144001
8       1024.0   399.904013   378.847986     537.728012   277.904004
9       2048.0   800.607979   767.359972    1053.311944   556.544006
10      4096.0  1619.567990  1522.143960    2200.223923  1127.392054
deepseek-ai/DeepSeek-V3 N=32768 K=512: 
fp8 blockwise scaled matmul:
    batch_size        vllm  sgl-kernel  sglang triton    deepgemm
0          1.0   16.416000   19.136000      15.232000   13.024000
1          8.0   16.031999   19.136000      13.760000   12.768000
2         16.0   16.384000   18.848000      13.344000   12.960000
3         32.0   16.384000   18.751999      13.152000   13.088000
4         64.0   16.160000   18.719999      13.376000   13.248000
5        128.0   16.096000   18.751999      18.271999   13.824000
6        256.0   21.312000   23.871999      28.255999   18.751999
7        512.0   34.784000   36.256000      46.239998   28.767999
8       1024.0   62.399998   61.439998      82.047999   46.944000
9       2048.0  116.159998  110.799998     153.408006   85.919999
10      4096.0  223.072007  207.519993     295.664012  165.184006
deepseek-ai/DeepSeek-V3 N=7168 K=16384: 
fp8 blockwise scaled matmul:
    batch_size         vllm   sgl-kernel  sglang triton    deepgemm
0          1.0    76.768003    78.960001     112.512000   60.224000
1          8.0    76.031998    78.432001      72.480001   57.856001
2         16.0    76.127999    78.624003      66.944003   57.087999
3         32.0    76.608002    78.720003      70.303999   50.080001
4         64.0    76.320000    79.039998      70.656002   41.887999
5        128.0    76.320000    78.720003      82.847998   53.247999
6        256.0    77.504002    80.416001     107.311994   58.143999
7        512.0   149.471998   150.624007     200.895995  117.919996
8       1024.0   289.696008   285.472006     471.136004  234.623998
9       2048.0   528.591990   522.607982     755.904019  485.760003
10      4096.0  1084.959984  1067.872047    1471.087933  733.471990
deepseek-ai/DeepSeek-V3 N=7168 K=18432: 
fp8 blockwise scaled matmul:
    batch_size         vllm   sgl-kernel  sglang triton    deepgemm
0          1.0    85.568003    88.096000     124.224000   65.215997
1          8.0    84.384002    87.007999      83.807997   63.423999
2         16.0    84.799998    86.176001      77.376001   61.567999
3         32.0    84.384002    86.400002      76.800004   54.400001
4         64.0    84.192000    86.592004      78.079998   45.759998
5        128.0    84.063999    86.687997      94.176002   55.936001
6        256.0    86.560003    89.120001     118.752003   66.143997
7        512.0   166.207999   167.104006     220.912009  124.191999
8       1024.0   348.863989   317.519993     514.559984  247.615993
9       2048.0   601.472020   588.096023     827.552021  517.664015
10      4096.0  1282.240033  1237.583995    1631.872058  824.751973
deepseek-ai/DeepSeek-V3 N=4608 K=7168: 
fp8 blockwise scaled matmul:
    batch_size        vllm  sgl-kernel  sglang triton    deepgemm
0          1.0   38.015999   41.407999      50.207999   29.023999
1          8.0   37.856001   40.768001      33.824001   26.912000
2         16.0   37.856001   40.768001      33.440001   25.312001
3         32.0   37.951998   40.895998      33.216000   21.919999
4         64.0   37.856001   40.959999      33.535998   19.808000
5        128.0   38.079999   41.120000      40.927999   21.824000
6        256.0   38.431998   41.536000      48.255999   26.752001
7        512.0   69.600001   71.584001      81.408001   36.607999
8       1024.0  101.375997  102.016002     135.263994   65.792002
9       2048.0  164.287999  160.607994     215.903997  114.271998
10      4096.0  299.392015  293.951988     409.103990  269.407988
deepseek-ai/DeepSeek-V3 N=3072 K=7168: 
fp8 blockwise scaled matmul:
    batch_size        vllm  sgl-kernel  sglang triton    deepgemm
0          1.0   37.567999   40.320002      41.855998   25.408000
1          8.0   37.471998   40.256001      29.503999   25.312001
2         16.0   37.503999   40.256001      29.279999   23.903999
3         32.0   37.535999   40.383998      29.023999   19.711999
4         64.0   37.471998   40.352002      29.216001   16.319999
5        128.0   37.696000   40.608000      32.000002   17.824000
6        256.0   37.728000   40.895998      40.128000   25.024001
7        512.0   38.304001   41.216001      48.928000   28.672000
8       1024.0   69.215998   71.392000      91.392003   41.664001
9       2048.0  101.807997  102.784000     136.255994   76.959997
10      4096.0  196.544006  198.016003     271.488011  132.927999
deepseek-ai/DeepSeek-V3 N=4096 K=512: 
fp8 blockwise scaled matmul:
    batch_size       vllm  sgl-kernel  sglang triton  deepgemm
0          1.0  10.272000   13.376000       9.312000     8.192
1          8.0  10.240000   13.248000       8.480000     7.936
2         16.0  10.272000   13.280000       8.288000     7.808
3         32.0  10.432000   13.280000       8.288000     8.000
4         64.0  10.272000   13.280000       8.160000     8.000
5        128.0  10.464000   13.296000       8.608000     8.032
6        256.0  10.384000   13.312000       9.504000     8.608
7        512.0  10.624000   13.504000      10.624000     9.792
8       1024.0  13.760000   16.287999      16.368000    12.128
9       2048.0  20.927999   23.040000      26.016001    17.600
10      4096.0  34.784000   36.384001      44.480000    27.424
deepseek-ai/DeepSeek-V3 N=3072 K=1536: 
fp8 blockwise scaled matmul:
    batch_size       vllm  sgl-kernel  sglang triton   deepgemm
0          1.0  14.528000   17.247999      14.336000  10.336000
1          8.0  14.656000   17.247999      12.032000  10.016000
2         16.0  14.560000   17.247999      11.584000   9.872000
3         32.0  14.528000   17.279999      11.744000   9.376000
4         64.0  14.592000   17.312000      11.680000   9.376000
5        128.0  14.656000   17.376000      12.320000   9.664000
6        256.0  14.816000   17.472001      13.792000  10.464000
7        512.0  14.976000   17.535999      16.480001  11.840000
8       1024.0  22.496000   24.831999      27.680000  16.287999
9       2048.0  30.463999   32.127999      38.015999  23.808001
10      4096.0  53.024001   54.912001      68.672001  37.632000
deepseek-ai/DeepSeek-V3 N=512 K=7168: 
fp8 blockwise scaled matmul:
    batch_size       vllm  sgl-kernel  sglang triton   deepgemm
0          1.0  20.608000   31.872001      31.615999  13.632000
1          8.0  20.608000   31.904001      26.912000  13.248000
2         16.0  20.703999   31.904001      26.944000  12.896000
3         32.0  20.768000   31.840000      26.784001  12.864000
4         64.0  20.752000   31.679999      26.912000  12.928000
5        128.0  20.736000   32.000002      26.784001  12.896000
6        256.0  21.056000   32.320000      26.784001  13.120000
7        512.0  21.792000   32.703999      27.295999  15.456000
8       1024.0  26.720000   31.840000      30.719999  16.448000
9       2048.0  33.535998   36.031999      39.551999  21.888001
10      4096.0  46.656001   44.000000      50.719999  29.983999
deepseek-ai/DeepSeek-V3 N=7168 K=2304: 
fp8 blockwise scaled matmul:
    batch_size        vllm  sgl-kernel  sglang triton    deepgemm
0          1.0   18.912001   21.919999      22.399999   14.528000
1          8.0   18.464001   21.376001      17.824000   14.176000
2         16.0   18.592000   21.504000      17.344000   14.048000
3         32.0   18.528000   21.407999      17.152000   13.696000
4         64.0   18.464001   21.504000      17.088000   13.312000
5        128.0   18.560000   21.632001      18.848000   14.720000
6        256.0   18.688001   21.792000      22.431999   16.416000
7        512.0   29.279999   31.583998      36.800001   24.224000
8       1024.0   50.624002   50.976001      77.312000   39.584000
9       2048.0   82.911998   80.991998     111.040004   71.199998
10      4096.0  158.976004  152.224004     218.096003  108.319998
deepseek-ai/DeepSeek-V3 N=7168 K=2048: 
fp8 blockwise scaled matmul:
    batch_size        vllm  sgl-kernel  sglang triton    deepgemm
0          1.0   17.696001   20.703999      21.760000   12.704000
1          8.0   17.216001   20.191999      14.816000   12.768000
2         16.0   17.600000   20.479999      14.624000   12.320000
3         32.0   17.408000   20.191999      14.592000   11.712000
4         64.0   17.376000   20.288000      14.688000   11.744000
5        128.0   17.535999   20.400001      16.543999   12.800000
6        256.0   17.535999   20.544000      20.096000   14.624000
7        512.0   26.912000   29.440001      33.087999   22.848001
8       1024.0   45.919999   46.688002      70.015997   38.240001
9       2048.0   75.456001   73.536001     101.888001   68.063997
10      4096.0  144.191995  137.503996     193.376005  102.080002
deepseek-ai/DeepSeek-V3 N=7168 K=256: 
fp8 blockwise scaled matmul:
    batch_size       vllm  sgl-kernel  sglang triton   deepgemm
0          1.0   9.440000   12.512000       8.736000   8.448000
1          8.0   9.408000   12.448000       7.616000   8.000000
2         16.0   9.408000   12.416000       7.584000   8.192000
3         32.0   9.504000   12.320000       7.584000   8.160000
4         64.0   9.504000   12.352000       7.904000   8.192000
5        128.0   9.408000   12.224000       8.480000   8.480000
6        256.0   9.760000   12.544000       9.568000   9.056000
7        512.0  11.904000   14.496000      12.992000  11.424000
8       1024.0  16.352000   18.848000      23.232000  15.104000
9       2048.0  23.808001   25.599999      34.272000  22.752000
10      4096.0  40.832002   41.824002      61.471999  31.615999
Benchmark finished!
```

pr:
```shell
DeepSeek-V3
tp=8

INFO 04-28 06:38:20 [__init__.py:239] Automatically detected platform cuda.
Skip N=576, K=7168 now
deepseek-ai/DeepSeek-V3 N=24576 K=7168: 
fp8 blockwise scaled matmul:
    batch_size         vllm   sgl-kernel  sglang triton     deepgemm
0          1.0    89.887999    92.224002      85.184000    76.416001
1          8.0    81.216000    83.456002      70.239998    70.592001
2         16.0    82.336001    84.608003      76.352000    65.056004
3         32.0    81.055999    83.071999      64.127997    59.583999
4         64.0    82.479998    84.352002      63.327998    57.952002
5        128.0    77.632003    79.935998      98.240003    61.360002
6        256.0   105.503999   106.112003     143.616006    86.687997
7        512.0   201.072007   197.392002     272.015989   136.656001
8       1024.0   395.135999   386.496007     536.880016   272.992015
9       2048.0   793.167949   762.272000    1054.816008   566.496015
10      4096.0  1594.303966  1581.055999    2187.151909  1143.615961
deepseek-ai/DeepSeek-V3 N=32768 K=512: 
fp8 blockwise scaled matmul:
    batch_size        vllm  sgl-kernel  sglang triton    deepgemm
0          1.0   16.480001   19.328000      15.840000   13.024000
1          8.0   16.128000   19.424001      13.696000   12.928000
2         16.0   15.968001   19.136000      13.536000   13.184000
3         32.0   16.224001   19.168001      13.280000   13.056000
4         64.0   16.000001   19.040000      13.728000   13.200000
5        128.0   16.128000   19.231999      18.400000   14.176000
6        256.0   21.536000   23.968000      28.287999   19.168001
7        512.0   34.880001   35.999998      46.271998   29.023999
8       1024.0   62.368002   60.256001      82.751997   47.488000
9       2048.0  117.215998  107.808001     155.328006   85.727997
10      4096.0  223.616004  199.647993     295.520008  165.408000
deepseek-ai/DeepSeek-V3 N=7168 K=16384: 
fp8 blockwise scaled matmul:
    batch_size         vllm   sgl-kernel  sglang triton    deepgemm
0          1.0    77.183999    76.991998     112.896003   60.384002
1          8.0    75.935997    76.640002      72.800003   58.240000
2         16.0    76.448001    76.672003      67.167997   57.023998
3         32.0    76.768003    76.831996      70.656002   50.239999
4         64.0    76.448001    77.151999      71.071997   41.855998
5        128.0    75.712003    76.448001      82.815997   53.376000
6        256.0    77.791996    78.272000     108.255997   58.432002
7        512.0   149.215996   147.264004     203.040004  118.207999
8       1024.0   289.664000   285.135984     474.687994  230.463997
9       2048.0   536.095977   511.615992     710.655987  466.239989
10      4096.0  1084.720016  1057.407975    1456.592083  710.687995
deepseek-ai/DeepSeek-V3 N=7168 K=18432: 
fp8 blockwise scaled matmul:
    batch_size         vllm   sgl-kernel  sglang triton    deepgemm
0          1.0    85.759997    85.663997     124.767996   65.600000
1          8.0    84.608003    85.023999      83.743997   63.840002
2         16.0    84.895998    84.927998      77.568002   61.535999
3         32.0    85.440002    85.120000      77.087998   54.623999
4         64.0    84.863998    84.991999      78.272000   46.176001
5        128.0    84.959999    85.199997      95.168002   55.872001
6        256.0    86.528003    87.583996     118.592001   66.367999
7        512.0   167.104006   164.031997     221.103996  124.512002
8       1024.0   329.216003   317.535996     513.599992  255.488008
9       2048.0   593.631983   577.744007     799.504042  493.696004
10      4096.0  1387.359977  1252.544045    1652.799964  782.815993
deepseek-ai/DeepSeek-V3 N=4608 K=7168: 
fp8 blockwise scaled matmul:
    batch_size        vllm  sgl-kernel  sglang triton    deepgemm
0          1.0   38.304001   40.031999      50.687999   28.928000
1          8.0   38.112000   39.744001      33.856001   26.815999
2         16.0   37.919998   39.744001      33.440001   25.728000
3         32.0   37.888002   39.840002      33.599999   22.112001
4         64.0   38.015999   39.840002      33.696000   20.000000
5        128.0   38.368002   40.031999      41.343998   22.016000
6        256.0   38.431998   40.640000      48.160002   26.880000
7        512.0   69.664001   69.343999      81.696004   36.543999
8       1024.0  101.760000   99.840000     135.008007   65.632001
9       2048.0  164.287999  161.200002     216.064006  114.784002
10      4096.0  298.752010  294.528008     406.816006  273.056000
deepseek-ai/DeepSeek-V3 N=3072 K=7168: 
fp8 blockwise scaled matmul:
    batch_size        vllm  sgl-kernel  sglang triton    deepgemm
0          1.0   37.664000   39.584000      42.048000   25.615999
1          8.0   37.439998   39.296001      29.696001   25.248000
2         16.0   37.728000   39.296001      29.503999   23.776000
3         32.0   37.503999   39.328001      29.408000   20.128001
4         64.0   37.567999   39.360002      29.600000   16.448000
5        128.0   37.983999   39.551999      32.032002   18.208001
6        256.0   38.112000   40.064000      40.320002   25.024001
7        512.0   38.527999   40.544000      48.928000   28.640000
8       1024.0   69.728002   70.271999      92.416003   41.664001
9       2048.0  102.112003  102.527998     136.575997   76.704003
10      4096.0  195.968002  193.599999     269.760013  133.184001
deepseek-ai/DeepSeek-V3 N=4096 K=512: 
fp8 blockwise scaled matmul:
    batch_size       vllm  sgl-kernel  sglang triton   deepgemm
0          1.0  10.432000   13.728000       9.696000   8.160000
1          8.0  10.560000   13.696000       8.704000   8.096000
2         16.0  10.432000   13.728000       8.544000   7.936000
3         32.0  10.656000   13.760000       8.480000   7.936000
4         64.0  10.720000   13.792000       8.352000   8.192000
5        128.0  10.464000   13.824000       8.800000   8.416000
6        256.0  10.528000   13.696000       9.696000   8.960000
7        512.0  10.912000   13.856000      11.104000   9.952000
8       1024.0  13.984000   16.767999      15.936000  12.256000
9       2048.0  21.215999   23.232000      25.888000  17.856000
10      4096.0  35.168000   35.872001      44.767998  27.712001
deepseek-ai/DeepSeek-V3 N=3072 K=1536: 
fp8 blockwise scaled matmul:
    batch_size       vllm  sgl-kernel  sglang triton   deepgemm
0          1.0  14.752000   17.472001      14.368000  10.544000
1          8.0  14.912000   17.472001      12.000000  10.432000
2         16.0  14.912000   17.472001      11.776000  10.112000
3         32.0  14.944000   17.503999      11.936000   9.536000
4         64.0  14.816000   17.519999      11.840000   9.600000
5        128.0  14.880000   17.600000      12.704000   9.664000
6        256.0  14.976000   17.728001      14.112000  10.400000
7        512.0  15.168000   17.792000      16.608000  11.968000
8       1024.0  22.720000   24.896000      27.840000  16.319999
9       2048.0  30.624000   32.352000      38.240001  24.224000
10      4096.0  53.952001   53.856000      69.343999  37.664000
deepseek-ai/DeepSeek-V3 N=512 K=7168: 
fp8 blockwise scaled matmul:
    batch_size       vllm  sgl-kernel  sglang triton   deepgemm
0          1.0  20.927999   31.599998      31.840000  13.824000
1          8.0  20.927999   31.615999      27.104000  13.472000
2         16.0  20.959999   31.583998      27.168000  13.248000
3         32.0  20.768000   31.552002      27.200000  13.184000
4         64.0  20.864001   31.328000      27.071999  13.152000
5        128.0  20.768000   31.711999      26.815999  13.056000
6        256.0  21.088000   32.000002      27.008001  13.472000
7        512.0  22.112001   32.288000      27.327999  16.096000
8       1024.0  26.848000   30.975999      30.880000  16.256001
9       2048.0  33.376001   34.432001      39.519999  21.663999
10      4096.0  46.751998   41.983999      50.912000  30.304000
deepseek-ai/DeepSeek-V3 N=7168 K=2304: 
fp8 blockwise scaled matmul:
    batch_size        vllm  sgl-kernel  sglang triton    deepgemm
0          1.0   19.200001   21.919999      22.399999   14.720000
1          8.0   18.448001   21.312000      17.824000   14.336000
2         16.0   18.624000   21.472000      17.535999   14.464000
3         32.0   18.560000   21.344000      17.152000   13.728000
4         64.0   18.719999   21.472000      17.472001   13.504000
5        128.0   18.688001   21.568000      19.040000   14.912000
6        256.0   19.072000   21.856001      22.431999   15.936000
7        512.0   29.536000   31.520002      37.280001   24.480000
8       1024.0   51.040001   51.584002      78.111999   39.935999
9       2048.0   83.296001   82.336001     111.040004   71.616001
10      4096.0  159.199998  155.103996     216.495991  109.151997
deepseek-ai/DeepSeek-V3 N=7168 K=2048: 
fp8 blockwise scaled matmul:
    batch_size        vllm  sgl-kernel  sglang triton    deepgemm
0          1.0   18.144000   20.768000      21.663999   13.056000
1          8.0   17.632000   20.288000      15.264000   12.832000
…