@zhyncs 2025-04-29
Hi @elfiegg https://github.com/sgl-project/sglang/blob/8d463fe351c4b10ed8b65d7fe5ed0c68eea6d731/sgl-kernel/benchmark/bench_fp8_blockwise_gemm.py#L115-L146