@BBuf 2025-04-29
> LGTM
> 
> QQ: What is the metric of the benchmark? I'm wondering if it's latency or throughput?

It is latency.